added 'all' directory and config.yml
