+++ recipes/cwt-cucumber/all/test_package/features/7_doc_strings.feature
