+++ recipes/cwt-cucumber/config.yml
@@ -1,0 +1,3 @@
+versions:
+  "2.5":
+    folder: all
